Implemented basic saving and loading
Created an IDataStorage interface that storage apis can implement allowing us to store and load data without caring about how an api works. Created a basic localStorage implementation for the IDataStorage interface. Stored the DataStorage on a Systems class so we can just call Systems.DataStorage as a static class and utilise the functionality. In the process of this, created an IGameState interface that holds everything that the game needs and is the stuff that will be saved and loaded.