Implemented basic functionality for updating attributes on turn change
When calling nextTurn() you can now supply an optional object that implements IAttributes, these will be the adjustments that will be made to the attributes in game state if passed in. Also implemented very basic StartScreen component that will be shown before the game is started and will have its continue button shown only if there is an existing save.