Refactored top-level App.tsx file
Split up the bulk into a Game.tsx component because it isn't really related to the overall App stateand instead uses the GameData from the save and just takes a nextTurn function and instead keeps its own state internally as a functional component. This is much easier to read and reason about as there is only 1 level of nesting in the App.tsx::render() function. Finally, also replaced all instances of 'region' with 'province' to keep everything consistent