Re-implement the Decision component
It is a lot cleaner now and follows the same modal styling as every other modal-like component. Should probably make those styles mixins. Also streamlined the IDecision interface. It now works on enums and then, when the functionality is in place, it will use those in conjunction with the current state of the nation to determine what affect to have on the overall nation attributes