Made the SVG pannable and zoomable on desktop
Experimented for ages and finally got the SVG map to be pannable and zoomable (via dragging and scrolling) on desktop. Made this into a component called PannableSVG that must have an SVG child nested inside of it, it will then clone that child and get a ref to it, returning the child so the SVG renders as normal but then using the ref it has obtained to add some event listeners, firstly it adds an onmousedown event listener which will enable dragging and store the mouse position when it is triggered, then it uses an onmousemove listener and, if the draggig is enabled it will update the SVGs viewBox by the movement values (varying by speed). Similar process for zooming but using the mousewheel event. Currently works as expected, although panning could be slightly smoother. Need to now repeat same functionality but for the the touch events