Final version for now
Both mouse and touch via pinch to zoom can control the map. currently it is quite laggy on mobile devices when using multi touch as adjusting the viewBox attribute on an SVG with this many elements is very slow and during pinch to zoom it does a looot of updates. I am happy to leave it as is for now because this only affects mobile devices and it only happens when zooming. However in the future it should probably be fixed